work in progress. Firebase works (connected) anyways. Just need to do visuals and iron everything out

+++ app/src/main/java/com/blacktweeter/android/twitter/data/FBCategory.java
@@ -9,7 +9,7 @@ import java.util.ArrayList;
 public class FBCategory {
     String name;
     String pictureUrl;
-    int orderNumber;
+    long orderNumber;
     ArrayList<FBTweet> tweetArray = new ArrayList<>();
 
     public FBCategory(){
@@ -31,7 +31,7 @@ public class FBCategory {
         this.pictureUrl = pictureUrl;
     }
 
-    public void setOrderNumber(int orderNumber) {
+    public void setOrderNumber(long orderNumber) {
         this.orderNumber = orderNumber;
     }
 
@@ -50,7 +50,7 @@ public class FBCategory {
         return pictureUrl;
     }
 
-    public int getOrderNumber() {
+    public long getOrderNumber() {
         return orderNumber;
     }
